requirements.txt
Django>=5.0,<6.0
Pillow>=10.0.0
qrcode>=7.4.2
mercadopago>=2.2.3
stripe>=11.0.0
python-decouple>=3.8
django-cleanup>=8.1.0
